// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ProblemFieldError: {
            field: string;
            message: string;
            code?: string | null;
        };
// ...
        Problem: {
            /** Format: uri */
            type: string;
            title: string;
            status: number;
            detail: string | null;
            code: string;
            /** Format: uri-reference */
            instance: string | null;
            traceId: string;
            errors: components["schemas"]["ProblemFieldError"][];
        };
// ...
    };
// ...
}